separate type declaration to distinct directory
